requirements.txt
pandas>=1.5.0
numpy>=1.21.0
yfinance>=0.2.0
ta-lib>=0.4.0
python-dotenv>=0.19.0
openai>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0 